Bug #1463: Fix issue changing order in parameters of entry point
The problem was that before the change, the children of the order returned a
proxy when you used the method getOrder() over them. However, this doesn't
happen if you change the order of the parameters.

I guess that this is because of the converter of the Order is now called before
the converter of the OrderElement and the problem disappears.

FEA: ItEr76S04BugFixing