Bug #1451: Error rendering GanttView coming directly from MonteCarlo
The error was in the code that builds the graph for calculating the
CriticalPath. There was a very similar code for building the graph in
TemplateModel, the only difference were the two lines of the patch.

These two lines were changed for fixing bug #1423. I tried this bug
again with the patch applied and it works too.

FEA: ItEr76S04BugFixing