Bug #1479: Fix error loading jqplot Javascript files
There was a syntax error loading the jquery javascript files that made
everything fail after.

"Unterminated string literal"

It seems that error is caused by not correctly closing the code that
embeds Javascript with a </script> tag, so I changed that part and fixed
the issue.

http://stackoverflow.com/questions/227552/common-sources-of-unterminated-string-literal

FEA: ItEr76S04BugFixing