Debian: Enable plpgsql when using PostgreSQL 8.x
In case PostgreSQL 8.x is being used, make sure that support
for plpgsql is enabled. This is needed starting with version
1.3 of Libreplan. Note that the major version number number
is taken from the output of "createlang --version", so this
assumes that the version for which the tool is available is
actually the one being run. This assumption should be safe
in the vast majority of cases.