Show only calculated indicators in earned value chart
In LibrePlan Audiovisual BCWP is not calculated as instead of standard progress
the filming progress perspective is used so there is not input data to calculate
BCWP. This causes that several indicators cannot be calculated either as they
depend on this value.

Only BCWS, ACWP and BAC will appear in the earned value chart.