Fix issue saving budget changes
Due to commit 7e79505c3e462718cbc5f47d870a240d84ccbd5d the changes in the budget
were not been saved.

So, the budget is only refreshed from database if it is already closed (the
order status is different from budget). Otherwise the budget associated to the
order was not been directly modified and the changes were not persisted into the
database.