Remove dependency to DejaVu fonts as are already included in jasperreport-fonts
Changes in commit 004dec3a429440374667defbe47f0be0ec6faa08 where not needed
because of the new maven dependency with jasperreport-fonts already includes the
DejaVu fonts. So we don't need to have any font installed in the system to
generate properly the reports.

That means that we will have one less dependency in the future versions.

FEA: ItEr77S08ReportsFonts