Bug #1537: Fix issue getting allocations from memory and not from database
Without this patch the generic allocations were gotten from database, and only
for the criteria in an already stored allocation a replacement for the
allocations in memory was done.

From now on, for each possible criterion, the allocations from database are
recovered, and they are replaced if needed by the new ones in memory.

FEA: ItEr77S04BugFixing