Initialize the start filtering date with the task start date and the end filtering date with the task end date.
It has been improved the performance so that the search of the resources and their
load ratios calculation is done from this moment on when the advanced search tab is made visible. This is
the moment when the task start date and task end date are available (not
sooner).

FEA: ItEr77S10ResourceAllocationLoadInformation