Add new button to adapt planning according to timesheets
It has been added a new method ICommand.isPlannerCommand() to define if a button
should be added in the planner toolbar or in the common toolbar (save and cancel
buttons).

For the moment, we are using a hard-coded value to know how many buttons we
should add in the plannerToolbar. At this moment we have 2 buttons: reassign and
adapt planning.

FEA: ItEr77S12AdaptPlanningAccordingTimesheets