Bug #1413: Fix bug
The methods used to calculate the start and end of the task when
moving the task and when allocating in the form were different. This
caused the length of the task to change when being moved.

Now the method used when moving the task is the one always in use,
because it's more precise.