Calculate progress and hours bars always proportionally to task size
Both bars will work like the money cost bar was already working.

The change is done due to the problem that appears when you adapt the planning,
and some planning remains before the start of the task. In that cases the bars
were not being printed properly.

Moreover current implementation was quite complex, and even if it gives some
useful information in specific cases, it also causes some misunderstandings to
the users.

FEA: ItEr77S12AdaptPlanningAccordingTimesheets