Added onClick listener to project names on company view to enter into the planning
This is done using the entry point /planner/index.zul;order=order-code which also helps
browser history management to enable bookmarking, refreshing or use the back button.

The link to the taks from the project view is not added due to an interaction it has with
the system to prevent from leaving the current URL, which would need to be written with anchors
to avoid this interaction.

FEA: ItEr77S03Community