Bug 1383: Fixed corner case of deadline and current day right position
The background position of the current day and deadline was outside the
visible area in the timetracker watermark columns when it was on the
last day of the interval.

A better alignment with the tasks layer (reducing the 3px max deviation)
on this cases can be achieved by revamping how lines and borders are
represented on the watermark layer.

FEA: ItEr76S04BugFixing