Adapted dates filter to user preferences
It will substract N months to the current date and use it as the start
date for the compan view filtering, being N the 'projects since'
parameter on the user settings tab.

Similarly, it will add M months to the current date and use it as
the filtering end date with the specified parameter.

FEA: ItEr77S15FilteringEnhancements