jira-integration: Move code part from updateWorkReportLine to updateOrCreateWorkReportLineAndAddToWorkReport
The code was being generated in 2 parts, first for the query and then to be set.
As in other cases if the WorkReportLine is found by code, we don't need to set
it again. So the part related to manage the code has been moved close to the
query and only set for newly created objects.

FEA: ItEr77S16JiraAndTimConnectorContributionIntegration