jira-integration: Avoid issue with code comming back from task edition pop-up
When you enter in the edition of a task in the WBS, if the task is a JIRA issue,
when you come back to the WBS LibrePlan was trying to update the code column,
however this column doesn't exist for JIRA issues as we're using a link.

Fixed that to avoid an exception in this situation.

FEA: ItEr77S16JiraAndTimConnectorContributionIntegration