Use new database query when filtering projects list
The database query that filters the orders by several fields is used only when
the subelements checkbox is not clicked. Otherwise the old query is used and the
old behavior is kept for the filtering.

However, if the checkbox is not clicked, the new query is used to get the
initial list and to do the successive filtering too.

NOTE: Resources are never assigned directly to the root element, so if you
filter by resources it'll only work if you have the subelements checkbox
clicked.

FEA: ItEr77S15FilteringEnhancements