Jira-integration: some improvement on get issues.
getIssues gets now the issues from jira in one request. This is done by adding a
query fields with desired fields to be included in response. Moreover the max_results
is now outside if(!query.isEmpty) statetment.
Test case is modified to read label from jira-properties instead of hard coded