Fix issue in resources load bandbox with value from session
The value from session was not been used to calculate the data to show, and it
was only renderer.

The problem was that the method
org.libreplan.web.resourceload.ResourceLoadController.Reloader.reloadInTransaction()
was calling
org.libreplan.web.resourceload.ResourceLoadController.WorkersOrCriteriaBandbox.setup(ResourcesLoadPanel)
after calculating the data to show.

So
org.libreplan.web.resourceload.ResourceLoadController.WorkersOrCriteriaBandbox.buildBandboxFilterer()
was called after it calculates the data to show, so it doesn't take into account
the default values for the selected filters.

Now, the initialization of the bandbox is done in the constructor, and then the
selected filters are marked.

Later the buildBandboxFilterer is called and it only wraps the bandbox inside a
hbox.

FEA: ItEr77S15FilteringEnhancements