Get zoom from session in resources load view
Add a listener in order to update the value in the session when user changes the
zoom.

For the company view it will use it's own variable to store the zoom level,
inside a project it will use the zoom variable shared for all the project views.

FEA: ItEr77S15FilteringEnhancements