Keep tasks converted into containers also in the filtered WBS
This allows to add a task over a leaf in a filtered WBS, and being able to keep
seeing the 3 elements: the new container (with the name of the previous leaf),
the previous leaf without name and the new task added.

It's easier to understand with an example. Imagine you have the following WBS:
* Task 1
* Task 2

And you filter by task name using "2", then you'll have:
* Task 2

Then you select "Task 2" and add a new task called "Subtask A".

Without this patch you'll only see:
* Task 2 (the new container using the name of the old leaf)
  |- Subtask A (the new task)

However, it's better you see the following
* Task 2 (the new container using the name of the old leaf)
  |- (empty task) (the old leaf that you should rename)
  |- Subtask A (the new task)

FEA: ItEr77S15FilteringEnhancements