Force synchronization of filter changes when switching between loaded perspectives
Either in the Company view/Projects list perspectives and the Project Gantt/WBS ones,
their filter values are synchronized when shitching between them. All the perspectives
were loading the rigth session values when they were loaded for the first time, but
needed to be updated after changes in the filter values to actualize the widgets
on the sibling perspective.

FEA: ItEr77S15FilteringEnhancements