tim-connector: Created new entity Connector based on AppProperties
AppProperties entity has been removed as it stores the same information
(majorId) repeated in each database record and it doesn't match with the real
model that we need.

A new entity Connector with a majorId identifier has been created, this entity
contains a list of properties (pairs key-value).

Moreover it has been created the predefined Tim connector with its own
properties (predefined too).

For new connectors, apart from the specific classes implementing them it would
be only needed to modify PredefinedConnectors and PredefinedConnectorProperties.
The database will be updated automatically on LibrePlan startup thanks to the
ConnectorBootstrap.

FEA: ItEr77S16JiraAndTimConnectorContributionIntegration