Bug #1563: Remove links from breadcrumbs component
The links on the breadcrumbs line were problematic as they
were pointing to the first subelement of the current section of the
menu, which could be an element unauthorized for the user.

As these shortcats are not useful enough and redundant with the main menu,
which can be used directly as is showing only the entries to the specific
user has access to, the links in this components are removed.

FEA: ItEr77S04BugFixing