Bug #1606: Revert fix for bug #1546 wich causes regressions on dependencies after saving project
This reverts commit 9a8d7deffe7bb5034448f698ab3dff4014b1bdb1.

The original fix on the patch "Bug #1546: Force recalculation of critical path progresses on
saving project" is having a side-effect on the behaviour of the dependencies, that don't push
tasks and are show violated.

As the regression effect is more severe than the original bug, we revert the patch while finding
another solution for the original issue.