Execute tests in alphabetical order
Until now the order of the execution of the tests depended on the
order of the files returned by the filesystem. This could lead to
different test results in different systems for the same codebase.

Ideally tests should be completely isolated and not depend on the
order of execution whatsoever. Nevertheless, in many tests some data
is left behind in the database, affecting the behavior of other tests.

With the alphabetical order, these order problems will arise the same
in all machines.