Fix issue with messages for users
The MessagesForUser component is used for showing information and
error messages to the users. It's designed to disappear once the user
does some action and a minimum of time has passed.

This is done listening to the events being produced in the page. In
previous versions of ZK this was enough to identify user activity, but
in the version currently used, 5.0.11, events not related to user
activity are generated. This causes the messages to disappear without
user action.

With this fix, only when the user clicks on something, or changes the
value of some input the shown message is cleared.