Fix performance regression
When opening the `resource allocation modal window` (double-click on a
task), the advanced search was calculating the load of the resources
when it should do it when clicking advanced search button. This leads
to a big delay when opening the `resource allocation modal window` if
there is a moderate amount of data.

Now these calculations are only done when clicking advanced search
button and the `resource allocation modal window` is opened almost
instantly.