Fix bug #1646
A regression was introduced when adding filtering for the WBS screen.
Instead of being associated to the original TreeModel, the tree is
associated to a newly created one based on the current filtering
predicate. So any changes to the original TreeModel aren't
automatically shown in the Tree UI.

A method must be called to associate a new TreeModel, with the new
changes incorporated, to the tree. The method `filterByPredicateIfAny`
implemented this functionality, but it was renamed to
`reloadTreeUIAfterChanges` to convey better its purpose.

Now when adding a new template, `reloadTreeUIAfterChanges` is called
and the newly created element is shown.